Axis map only use data connected to the given axis.
Updated `getAxisMapByAxes` to only use the data sets connected to the
given axis, avoiding some _empty_ points from being added to the chart's
axis.

In my case I had some Line and Scatter components together but they have
different data sets and with different intervals. With this update I can
have the Line's axis displaying only the numbers and categories related
to it and the same happens to the Scatter's axis.