Factor out some scale- and rect-related functions (#5)
Goal: Make it easy to allow fixed endpoints in ReferenceLine

Create a `ScaleHelper` to provide two useful helper functions on top of what d3-scale provides, and a `LabeledScaleHelper` which makes it convenient to work with multiple scales at once. Includes unit tests of those helpers.

Refactor the reference areas to use the new helpers.